server/requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.1.0
openpyxl==3.1.2
xlrd==2.0.1
weasyprint==60.1
python-multipart==0.0.22
